restructure navbar
