+++ src/components/Navbar/SelectionBar.tsx
@@ -1,0 +1,6 @@
+import NavbarBase from 'components/Navbar/base';
+import styled from 'styled-components';
+
+export const SelectionBar = styled(NavbarBase)`
+    position: fixed;
+`;
