+++ src/components/pages/gallery/SelectedFileOptions/index.tsx
@@ -1,4 +1,4 @@
-import Navbar from 'components/Navbar';
+import Navbar from 'components/Navbar/base';
 import styled from 'styled-components';
 
 export const SelectionBar = styled(Navbar)`
